# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI GO MALL 后台 API 文档（仅 dev 环境暴露）。",
        "title": "ai-go-mall API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/login": {
            "post": {
                "description": "校验点选验证码 → 校验账号 → 签发 token；前端拿到 token 后自行保存并放入后续请求头。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录请求（含用户名 / 密码 / 点选坐标）",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_admin.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "login.invalid_input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "login.invalid_credentials / login.invalid_captcha",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "login.account_disabled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "login.internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/common/captcha/create": {
            "get": {
                "description": "服务端生成一道点选验证码；返回的 key 须先经 /common/captcha/verify 预校验坐标精度，才能用于业务（如登录）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "生成点选验证码",
                "responses": {
                    "200": {
                        "description": "成功：返回 key / elements / image 等",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_captcha.CreateClickResp"
                        }
                    },
                    "500": {
                        "description": "captcha.internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/common/captcha/verify": {
            "post": {
                "description": "仅校验坐标精度，不消耗 key（deleteOnSuccess=false）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "校验点选坐标精度",
                "parameters": [
                    {
                        "description": "点选坐标 + 验证码 key",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_captcha.VerifyClickReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功：{ ok: true }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "captcha.invalid_input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "captcha.mismatch",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "captcha.not_found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "410": {
                        "description": "captcha.expired",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "captcha.internal",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handler_admin.CaptchaPoint": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_admin.LoginRequest": {
            "type": "object",
            "required": [
                "captcha_key",
                "password",
                "points",
                "username"
            ],
            "properties": {
                "captcha_key": {
                    "description": "CaptchaKey 来源于点选验证码预检通过后回调透传的 key；\n缺 / 空 走 400，service 层做二次校验（consume 语义）。",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "points": {
                    "description": "Points 是用户在图片坐标系下的点击坐标（按 elements 顺序）。\n与 captchaKey 一起透传给 service，做二次校验。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_admin.CaptchaPoint"
                    }
                },
                "remember": {
                    "description": "Remember 记住我标记；true 表示 token 有效期 30 天，false（缺省）表示 3 天。\n前端不传时 Go 零值是 false，无需 omitempty。",
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_handler_admin.LoginResponse": {
            "type": "object",
            "properties": {
                "admin": {
                    "$ref": "#/definitions/internal_handler_admin.adminInfo"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_handler_admin.adminInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_handler_captcha.CreateClickResp": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_captcha.Point": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "internal_handler_captcha.VerifyClickReq": {
            "type": "object",
            "required": [
                "h",
                "key",
                "points",
                "w"
            ],
            "properties": {
                "h": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_captcha.Point"
                    }
                },
                "w": {
                    "type": "integer"
                }
            }
        }
    }
}